refactor(amqlib): split connect to client and server, use fanout and rpc_queue
