+++ microservices/amqlib/connect.go
@@ -2,8 +2,6 @@ package amqlib
 
 import (
 	"context"
-	"fmt"
-	"reflect"
 	"time"
 
 	"github.com/rabbitmq/amqp091-go"
@@ -22,347 +20,4 @@ type Connect struct {
 	Context context.Context
 	config  microservices.Config
 	timeout time.Duration
-}
-
-// Client usage
-func NewClient(opt Options) microservices.ClientProxy {
-	conn, err := amqp091.Dial(opt.Addr)
-	if err != nil {
-		fmt.Printf("Failed to connect to RabbitMQ: %v\n", err)
-		panic(err)
-	}
-
-	fmt.Printf("Connected to RabbitMQ at %s\n", opt.Addr)
-	connect := &Connect{
-		Context: context.Background(),
-		Conn:    conn,
-		config:  microservices.NewConfig(opt.Config),
-	}
-
-	return connect
-}
-
-func (c *Connect) Config() microservices.Config {
-	return c.config
-}
-
-func (c *Connect) Emit(event string, message microservices.Message) error {
-	return nil
-}
-
-func (c *Connect) Timeout(duration time.Duration) microservices.ClientProxy {
-	c.timeout = duration
-	return c
-}
-
-func (c *Connect) Send(event string, data any, res any, headers ...microservices.Header) error {
-	defer c.Conn.Close()
-
-	ch, err := c.Conn.Channel()
-	if err != nil {
-		return err
-	}
-	defer ch.Close()
-
-	err = ch.ExchangeDeclare(
-		"logs_direct", // name
-		"direct",      // type
-		true,          // durable
-		false,         // auto-deleted
-		false,         // internal
-		false,         // no-wait
-		nil,           // arguments
-	)
-	if err != nil {
-		return err
-	}
-
-	if c.timeout == 0 {
-		c.timeout = 5 * time.Second
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), c.timeout)
-	defer cancel()
-
-	body, err := microservices.EncodeMessage(c, microservices.Message{
-		Event:   event,
-		Data:    data,
-		Headers: microservices.AssignHeader(c.config.Header, headers...),
-	})
-	if err != nil {
-		return err
-	}
-	err = ch.PublishWithContext(ctx,
-		"logs_direct", // exchange
-		event,         // routing key
-		false,         // mandatory
-		false,         // immediate
-		amqp091.Publishing{
-			ContentType: "text/plain",
-			Body:        body,
-		})
-	if err != nil {
-		return err
-	}
-
-	c.timeout = 0
-	fmt.Printf("Send message: %v for event %s\n", data, event)
-	return nil
-}
-
-func (c *Connect) Publish(event string, data interface{}, headers ...microservices.Header) error {
-	defer c.Conn.Close()
-
-	ch, err := c.Conn.Channel()
-	if err != nil {
-		return err
-	}
-	defer ch.Close()
-
-	err = ch.ExchangeDeclare(
-		"logs_direct", // name
-		"direct",      // type
-		true,          // durable
-		false,         // auto-deleted
-		false,         // internal
-		false,         // no-wait
-		nil,           // arguments
-	)
-	if err != nil {
-		return err
-	}
-
-	if c.timeout == 0 {
-		c.timeout = 5 * time.Second
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), c.timeout)
-	defer cancel()
-
-	body, err := microservices.EncodeMessage(c, microservices.Message{
-		Event:   event,
-		Data:    data,
-		Headers: microservices.AssignHeader(c.config.Header, headers...),
-	})
-	if err != nil {
-		return err
-	}
-	err = ch.PublishWithContext(ctx,
-		"logs_direct", // exchange
-		event,         // routing key
-		false,         // mandatory
-		false,         // immediate
-		amqp091.Publishing{
-			ContentType: "text/plain",
-			Body:        body,
-		})
-	if err != nil {
-		return err
-	}
-
-	c.timeout = 0
-	return nil
-}
-
-// Server usage
-func New(module core.ModuleParam, opts ...Options) core.Service {
-	connect := &Connect{
-		Module:  module(),
-		config:  microservices.DefaultConfig(),
-		Context: context.Background(),
-	}
-
-	if len(opts) > 0 {
-		if opts[0].Addr != "" {
-			conn, err := amqp091.Dial(opts[0].Addr)
-			if err != nil {
-				panic(err)
-			}
-			connect.Conn = conn
-		}
-		if !reflect.ValueOf(opts[0].Config).IsZero() {
-			connect.config = microservices.ParseConfig(opts[0].Config)
-		}
-	}
-
-	return connect
-}
-
-func Open(opts ...Options) core.Service {
-	connect := &Connect{
-		config:  microservices.DefaultConfig(),
-		Context: context.Background(),
-	}
-
-	if len(opts) > 0 {
-		if opts[0].Addr != "" {
-			conn, err := amqp091.Dial(opts[0].Addr)
-			if err != nil {
-				panic(err)
-			}
-			connect.Conn = conn
-			fmt.Printf("Connected to RabbitMQ at %s\n", opts[0].Addr)
-		}
-		if !reflect.ValueOf(opts[0].Config).IsZero() {
-			connect.config = microservices.ParseConfig(opts[0].Config)
-		}
-	}
-
-	return connect
-}
-
-func (c *Connect) Create(module core.Module) {
-	c.Module = module
-}
-
-func (c *Connect) Listen() {
-	fmt.Println("Listening to RabbitMQ")
-	store := c.Module.Ref(microservices.STORE).(*microservices.Store)
-	if store == nil {
-		panic("store not found")
-	}
-
-	ch, err := c.Conn.Channel()
-	if err != nil {
-		panic(err)
-	}
-
-	err = ch.ExchangeDeclare(
-		"logs_direct", // name
-		"direct",      // type
-		true,          // durable
-		false,         // auto-deleted
-		false,         // internal
-		false,         // no-wait
-		nil,           // arguments
-	)
-	if err != nil {
-		panic(err)
-	}
-
-	if store.Subscribers != nil {
-		for _, sub := range store.Subscribers {
-			q, err := ch.QueueDeclare(
-				"",    // name
-				false, // durable
-				false, // delete when unused
-				true,  // exclusive
-				false, // no-wait
-				nil,   // arguments
-			)
-			if err != nil {
-				panic(err)
-			}
-
-			err = ch.QueueBind(
-				q.Name,        // queue name
-				sub.Name,      // routing key
-				"logs_direct", // exchange
-				false,
-				nil)
-			if err != nil {
-				panic(err)
-			}
-			go c.HandlerPubSub(ch, q.Name, sub)
-		}
-	}
-
-	if store.RpcHandlers != nil {
-		for _, sub := range store.RpcHandlers {
-			q, err := ch.QueueDeclare(
-				"",    // name
-				false, // durable
-				false, // delete when unused
-				true,  // exclusive
-				false, // no-wait
-				nil,   // arguments
-			)
-			if err != nil {
-				panic(err)
-			}
-			err = ch.QueueBind(
-				q.Name,        // queue name
-				sub.Name,      // routing key
-				"logs_direct", // exchange
-				false,
-				nil)
-			if err != nil {
-				panic(err)
-			}
-			go c.HandlerRPC(ch, q.Name, sub)
-		}
-	}
-}
-
-func (c *Connect) HandlerPubSub(ch *amqp091.Channel, q string, sub *microservices.SubscribeHandler) {
-	msgs, err := ch.Consume(
-		q,     // queue
-		"",    // consumer
-		true,  // auto-ack
-		false, // exclusive
-		false, // no-local
-		false, // no-wait
-		nil,   // args
-	)
-	if err != nil {
-		fmt.Printf("Error consuming messages: %v\n", err)
-		return
-	}
-	for d := range msgs {
-		message := microservices.DecodeMessage(c, d.Body)
-		fmt.Printf("Received message: %v for event %s\n", message.Data, message.Event)
-		if reflect.ValueOf(message).IsZero() {
-			sub.Handle(c, microservices.Message{
-				Bytes: d.Body,
-			})
-		} else {
-			sub.Handle(c, message)
-		}
-	}
-}
-
-func (c *Connect) HandlerRPC(ch *amqp091.Channel, q string, sub *microservices.RpcHandler) {
-	msgs, err := ch.Consume(
-		q,     // queue
-		"",    // consumer
-		true,  // auto-ack
-		false, // exclusive
-		false, // no-local
-		false, // no-wait
-		nil,   // args
-	)
-	if err != nil {
-		return
-	}
-	for d := range msgs {
-		message := microservices.DecodeMessage(c, d.Body)
-		var ctx microservices.Ctx
-		if reflect.ValueOf(message).IsZero() {
-			ctx = microservices.NewCtx(microservices.Message{
-				Data: d.Body,
-			}, c)
-		} else {
-			ctx = microservices.NewCtx(message, c)
-		}
-
-		reply, err := sub.Factory(ctx)
-		if err != nil {
-			ctx.ErrorHandler(err)
-			continue
-		}
-
-		if d.ReplyTo != "" {
-			err = ch.PublishWithContext(c.Context,
-				"",        // exchange
-				d.ReplyTo, // routing key
-				false,     // mandatory
-				false,     // immediate
-				amqp091.Publishing{
-					ContentType:   "application/json",
-					CorrelationId: d.CorrelationId,
-					Body:          reply,
-				})
-			if err != nil {
-				ctx.ErrorHandler(err)
-			}
-		}
-	}
 }
